# specs/003-dashboard-fix/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Dashboard Application API
  description: API for the dashboard application with task management, projects, tags, and notifications
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.yourapp.com
    description: Production server
paths:
  /api/v1/auth/login:
    post:
      summary: User login
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                  description: User's email address
                password:
                  type: string
                  description: User's password
              required:
                - username
                - password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: bearer
        '401':
          description: Invalid credentials
  /api/v1/auth/register:
    post:
      summary: User registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                password:
                  type: string
                  minLength: 8
              required:
                - email
                - name
                - password
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input or user already exists
...
  /api/v1/tasks:
    get:
      summary: Get all user tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    post:
      summary: Create a new task
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '200':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Invalid input
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [pending, in-progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
        userId:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - priority
        - userId
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        status:
          type: string
          enum: [pending, in-progress, completed]
        priority:
          type: string
          enum: [low, medium, high]
        dueDate:
          type: string
          format: date-time
        projectId:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - status
        - priority
...